+ Rename userspace radio files.
