+++ kernel/nrf/nrf_rf.c
@@ -9,9 +9,8 @@
 #include "nrf_rf.h"
 #include "nrf_power.h"
 #include "../../userspace/sys.h"
-#include "../../userspace/nrf/radio.h"
+#include "../../userspace/nrf/nrf_radio.h"
 #include "../../userspace/nrf/nrf_driver.h"
-#include "../../userspace/nrf/radio_config.h"
 #include "../kerror.h"
 #include "../kstdlib.h"
 #include "../ksystime.h"
